Add Module to SER
Increases version to v.0.4.2
