+++ Server/ModuleSER1/src/main/java/org/ModuleSER1/Domain/Assignment.java
@@ -1,20 +1,20 @@
-package A02.Server.Domain;
+package org.ModuleSER1.Domain;
 
 /** 
  * <p>
  *         <b>Class:</b></br>
- *         Assignment_Server_DAO.
+ *         Assignment.
  * </p>
  * 
  * <p>
  *         <b>Description:</b></br>
- *         Data Access Object of the Assignment.
+ *         Class that represents an Assignment.
  * </p>
  * 
  * <p>
  *         <b>Control change:</b></br>
  *         <ul>
- *             <li> <b>0.1.0:</b> Add methods add and read.
+ *             <li> <b>0.1.0:</b> Create class.
  *         </ul>
  * </p>
  * 
@@ -22,13 +22,16 @@ package A02.Server.Domain;
  * @since 0.1.0
  * @version 0.1.0
  */
-public class Assignment_Server_DAO {
+public class Assignment {
+	private int _assignmentID;
+	private String _memberID;
+	private String _courseID;
 
-	public void add(Attendance aAssignment) {
+	public Assignment(int aAssID, String aMember, int aCourse) {
 		throw new UnsupportedOperationException();
 	}
 
-	public Attendance read(int aAssID) {
+	public String toString() {
 		throw new UnsupportedOperationException();
 	}
 }
