+++ Server/ModuleSER1/src/main/java/org/ModuleSER1/Domain/Event_Types.java
@@ -1,4 +1,4 @@
-package A02.Server.Domain;
+package org.ModuleSER1.Domain;
 
 public enum Event_Types {
 	_teaching, _divulgative,_informational,_management;
